[CI] Add clang job and setup concurrency (#1407)

+++ steering_controllers_library/src/steering_controllers_library.cpp
@@ -85,7 +85,8 @@ controller_interface::CallbackReturn SteeringControllersLibrary::on_configure(
   const rclcpp_lifecycle::State & /*previous_state*/)
 {
   params_ = param_listener_->get_params();
-  odometry_.set_velocity_rolling_window_size(params_.velocity_rolling_window_size);
+  odometry_.set_velocity_rolling_window_size(
+    static_cast<size_t>(params_.velocity_rolling_window_size));
 
   configure_odometry();
